+++ tests/source/issue-3132.rs
@@ -1,0 +1,13 @@
+fn test() {
+    /*
+    a
+	*/
+    let x = 42;
+    /*
+    aaa
+    "line 1
+  line 2
+        line 3"
+    */
+    let x = 42;
+}
